+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -14,7 +14,7 @@ import { Montpellier } from './20221206_Montpellier';
 import { Mrn } from './20220420_MetropoleRouenNormandie';
 import { NantesMetropole2024 } from './20240101_NantesMetropole';
 import { Nm } from './20211202_NantesMetropoleXP';
-import { Occitanie20232024 } from './Occitanie20232024';
+import { Occitanie20232024 } from './20221024_Occitanie';
 import { PMGFxATMB2024 } from './PMGFxATMB2024';
 import { PaysBasque20232024 } from './PaysBasque20232024';
 import { PaysDeLaLoire2021 } from './PaysDeLaLoire2021';
